requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
urllib3==1.26.12
